+++ CHANGES.rst
@@ -24,6 +24,8 @@ Changelog
   [chris-adam]
 - Added validator on person's userid.
   [chris-adam]
+- Defined new Content Category Groups for dms files and appendix files.
+  [chris-adam]
 
 3.0 (2021-09-30)
 ----------------